Implemented screening, formatted code

+++ final/ticket-service/src/main/java/com/epam/training/ticketservice/core/movies/Movie.java
@@ -2,7 +2,6 @@ package com.epam.training.ticketservice.core.movies;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -30,7 +29,7 @@ public class Movie {
 
     @Id
     @GeneratedValue
-    private int Id;
+    private Long Id;
 
     @Column(unique = true)
     private String title;
